requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.2
kombu==5.4.2
prompt_toolkit==3.0.48
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
yt-dlp==2024.11.4
